requirements.txt
fastapi
uvicorn
pandas
scikit-learn
scipy
numpy
python-multipart
pydantic
joblib
pymongo
dnspython
python-dotenv
